EduLeaf/ml-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
PyPDF2>=3.0.1
python-docx>=1.1.2
Pillow>=10.3.0
pytesseract>=0.3.10
pydantic>=2.6.4
pydantic-settings>=2.2.1
httpx>=0.27.0
